ForgotPasswordResponse model added, some code refactors
